requirements.txt
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
langchain-community
pypdf
openai
celery
redis
sqlalchemy